feat: card edit functionality

+++ src/pages/api/cards/[id]/index.ts
@@ -1,10 +1,10 @@
-import { NextApiRequest, NextApiResponse } from 'next';
+import type { NextApiRequest, NextApiResponse } from 'next';
 import { Firestore } from '@google-cloud/firestore';
 import { NextStepsCard } from '#/types/new-here';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   const { id } = req.query;
-  if (req.method !== 'DELETE') {
+  if (req.method !== 'DELETE' && req.method !== 'PUT') {
     res.status(404).end();
   }
 
@@ -12,8 +12,24 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     projectId: 'next-steps-350612',
   });
 
-  await db.doc(`/cards/${id}`).delete();
-  res.status(200).end();
+  if (req.method === 'DELETE') {
+    await db.doc(`/cards/${id}`).delete();
+    res.status(200).end();
+    return;
+  }
+
+  if (req.method === 'PUT') {
+    const body = JSON.parse(req.body);
+    body.dob = new Date(body.dob);
+    body.date = new Date(body.date);
+
+    const doc = db.collection('cards').doc(id as string);
+    await doc.update(body);
+    res.status(200).end();
+    return;
+  }
+
+  res.status(418).end();
 };
 
 export default handler;
